requirements.txt
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
SQLAlchemy==1.2.18
Werkzeug==0.14.1
flask-cors
pyjwt

